renames slow hamiltonian for later reference
